+++ test/useridmapping/createUserIdMapping.test.js
@@ -5,6 +5,7 @@ const STExpress = require("../..");
 const { ProcessState } = require("../../lib/build/processState");
 const EmailPasswordRecipe = require("../../lib/build/recipe/emailpassword").default;
 const SessionRecipe = require("../../lib/build/recipe/session").default;
+const UserMetadataRecipe = require("../../lib/build/recipe/usermetadata").default;
 const { Querier } = require("../../lib/build/querier");
 const { maxVersion } = require("../../lib/build/utils");
 
@@ -179,6 +180,90 @@ describe(`createUserIdMappingTest: ${printPath("[test/useridmapping/createUserId
                 assert.strictEqual(createUserIdMappingResponse.status, "USER_ID_MAPPING_ALREADY_EXISTS_ERROR");
                 assert.strictEqual(createUserIdMappingResponse.doesSuperTokensUserIdExist, false);
                 assert.strictEqual(createUserIdMappingResponse.doesExternalUserIdExist, true);
+            }
+        });
+
+        it("create a userId mapping when userId already has usermetadata with and without force", async function () {
+            await startST();
+
+            STExpress.init({
+                supertokens: {
+                    connectionURI: "http://localhost:8080",
+                },
+                appInfo: {
+                    apiDomain: "api.supertokens.io",
+                    appName: "SuperTokens",
+                    websiteDomain: "supertokens.io",
+                },
+                recipeList: [EmailPasswordRecipe.init(), UserMetadataRecipe.init(), SessionRecipe.init()],
+            });
+
+            // Only run for version >= 2.15
+            let querier = Querier.getNewInstanceOrThrowError(undefined);
+            let apiVersion = await querier.getAPIVersion();
+            if (maxVersion(apiVersion, "2.14") === "2.14") {
+                return this.skip();
+            }
+
+            // create a user
+
+            const signInResponse = await EmailPasswordRecipe.signUp("test@example.com", "testPass123");
+            assert.strictEqual(signInResponse.status, "OK");
+            const superTokensUserId = signInResponse.user.id;
+
+            // add metadata to the user
+            const testMetadata = {
+                role: "admin",
+            };
+            await UserMetadataRecipe.updateUserMetadata(superTokensUserId, testMetadata);
+
+            const externalId = "externalId";
+            // without force
+            {
+                try {
+                    await STExpress.createUserIdMapping({
+                        superTokensUserId,
+                        externalUserId: externalId,
+                    });
+                    throw new Error("Should not come here");
+                } catch (error) {
+                    assert(error.message.includes("UserId is already in use in UserMetadata recipe"));
+                }
+            }
+            // with force set to false
+            {
+                try {
+                    await STExpress.createUserIdMapping({
+                        superTokensUserId,
+                        externalUserId: externalId,
+                        force: false,
+                    });
+                    throw new Error("Should not come here");
+                } catch (error) {
+                    assert(error.message.includes("UserId is already in use in UserMetadata recipe"));
+                }
+            }
+            // with force set to true
+            {
+                {
+                    let response = await STExpress.createUserIdMapping({
+                        superTokensUserId,
+                        externalUserId: externalId,
+                        force: true,
+                    });
+                    assert.strictEqual(response.status, "OK");
+                }
+
+                // check that mapping exists
+                {
+                    let response = await STExpress.getUserIdMapping({
+                        userId: superTokensUserId,
+                        userIdType: "SUPERTOKENS",
+                    });
+                    assert.strictEqual(response.status, "OK");
+                    assert.strictEqual(response.superTokensUserId, superTokensUserId);
+                    assert.strictEqual(response.externalUserId, externalId);
+                }
             }
         });
     });